+++ go/pkg/utils/utils.go
@@ -2,11 +2,22 @@ package utils
 
 import (
 	"bytes"
+	"context"
 	"fmt"
+	"io"
 	"log"
+	"os"
 	"os/exec"
 	"strings"
+	"time"
 
+	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
+	"k8s.io/apimachinery/pkg/util/yaml"
+	"k8s.io/apimachinery/pkg/watch"
+	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 )
@@ -39,4 +50,95 @@ func GetCurrentContextName(config *rest.Config, kubeconfigPath string) (string, 
 	}
 
 	return currentContext, nil
+}
+
+// ApplyManifestsFile applies all manifests in a provided file
+func ApplyManifestsFile(dynamicClient dynamic.Interface, manifestFile string) error {
+	fileData, err := os.ReadFile(manifestFile)
+	if err != nil {
+		return fmt.Errorf("failed to read %s: %w", manifestFile, err)
+	}
+
+	decoder := yaml.NewYAMLOrJSONDecoder(bytes.NewReader(fileData), 4096)
+	for {
+		var obj unstructured.Unstructured
+		if err := decoder.Decode(&obj); err != nil {
+			if err == io.EOF {
+				break // End of file, exit the loop
+			}
+			return fmt.Errorf("failed to decode YAML document: %w", err)
+		}
+
+		if obj.Object == nil {
+			continue // Skip empty objects
+		}
+
+		gvr := schema.GroupVersionResource{
+			Group:    obj.GetObjectKind().GroupVersionKind().Group,
+			Version:  obj.GetObjectKind().GroupVersionKind().Version,
+			Resource: getResourceName(obj.GetKind()),
+		}
+
+		_, err := dynamicClient.Resource(gvr).Namespace(obj.GetNamespace()).Create(context.TODO(), &obj, metav1.CreateOptions{})
+		if err != nil {
+			return fmt.Errorf("failed to apply resource (Kind: %s, Name: %s): %w", obj.GetKind(), obj.GetName(), err)
+		}
+	}
+
+	return nil
+}
+
+func getResourceName(kind string) string {
+	// Handle special cases
+	specialCases := map[string]string{
+		"Endpoints":                "endpoints",
+		"PodSecurityPolicy":        "podsecuritypolicies",
+		"NetworkPolicy":            "networkpolicies",
+		"CustomResourceDefinition": "customresourcedefinitions",
+	}
+
+	if resourceName, ok := specialCases[kind]; ok {
+		return resourceName
+	}
+
+	// Default case: add an 's' to the kind
+	return strings.ToLower(kind) + "s"
+}
+
+func WaitForCRDs(dynamicClient dynamic.Interface, crds []string) error {
+	for _, crd := range crds {
+		if err := waitUntilCRDEstablished(dynamicClient, crd); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+func waitUntilCRDEstablished(dynamicClient dynamic.Interface, crdName string) error {
+	w, err := dynamicClient.Resource(schema.GroupVersionResource{Group: "apiextensions.k8s.io", Version: "v1", Resource: "customresourcedefinitions"}).Watch(context.TODO(), metav1.ListOptions{
+		FieldSelector: "metadata.name=" + crdName,
+	})
+	if err != nil {
+		return fmt.Errorf("failed to watch CRD %s: %w", crdName, err)
+	}
+	defer w.Stop()
+
+	timeout := time.After(5 * time.Minute)
+	for {
+		select {
+		case event := <-w.ResultChan():
+			log.Printf("In CHANNEL %s\n", event.Type)
+			if event.Type == watch.Modified {
+				crd := event.Object.(*apiextensionsv1.CustomResourceDefinition)
+				log.Println("In watch")
+				for _, cond := range crd.Status.Conditions {
+					if cond.Type == apiextensionsv1.Established && cond.Status == apiextensionsv1.ConditionTrue {
+						return nil
+					}
+				}
+			}
+		case <-timeout:
+			return fmt.Errorf("timeout waiting for CRD %s to be established", crdName)
+		}
+	}
 }
